fix(DTFS2-6621): add tests
